ScanConfig to exclude directories from scanning.
